+++ src/main/java/roomescape/controller/ReservationTimeController.java
@@ -13,11 +13,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.request.ReservationTimeRequest;
-import roomescape.dto.request.ReservationTimeWithBookStatusRequest;
-import roomescape.dto.response.ReservationTimeResponse;
-import roomescape.dto.response.ReservationTimeWithBookStatusResponse;
 import roomescape.service.ReservationTimeService;
+import roomescape.service.dto.request.ReservationTimeRequest;
+import roomescape.service.dto.request.ReservationTimeWithBookStatusRequest;
+import roomescape.service.dto.response.ReservationTimeResponse;
+import roomescape.service.dto.response.ReservationTimeWithBookStatusResponse;
 
 @RestController
 @RequestMapping("/times")
