+++ src/main/java/roomescape/controller/RoomThemeController.java
@@ -10,9 +10,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.request.RoomThemeCreateRequest;
-import roomescape.dto.response.RoomThemeResponse;
 import roomescape.service.RoomThemeService;
+import roomescape.service.dto.request.RoomThemeCreateRequest;
+import roomescape.service.dto.response.RoomThemeResponse;
 
 @RestController
 @RequestMapping("/themes")
