+++ src/main/java/roomescape/handler/GlobalExceptionHandler.java
@@ -4,6 +4,7 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.handler.dto.ExceptionResponse;
 
 @RestControllerAdvice
 public class GlobalExceptionHandler {
