+++ src/main/java/roomescape/handler/dto/ExceptionResponse.java
@@ -1,4 +1,4 @@
-package roomescape.handler;
+package roomescape.handler.dto;
 
 import org.springframework.http.HttpStatus;
