+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -8,11 +8,11 @@ import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.domain.Reservation;
 import roomescape.domain.ReservationTime;
-import roomescape.dto.request.ReservationTimeRequest;
-import roomescape.dto.request.ReservationTimeWithBookStatusRequest;
-import roomescape.dto.response.ReservationTimeResponse;
-import roomescape.dto.response.ReservationTimeWithBookStatusResponse;
 import roomescape.handler.BadRequestException;
+import roomescape.service.dto.request.ReservationTimeRequest;
+import roomescape.service.dto.request.ReservationTimeWithBookStatusRequest;
+import roomescape.service.dto.response.ReservationTimeResponse;
+import roomescape.service.dto.response.ReservationTimeWithBookStatusResponse;
 
 @Service
 public class ReservationTimeService {
